requirements.txt
Pillow>=12.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
numpy

# For iOS Support
requests>=2.31.0

# For Model Deployment

## After installing sglang or vLLM, please run pip install -U transformers again to upgrade to 5.0.0rc0.
## Any dependency conflicts related to Transformers can be ignored.

# sglang>=0.5.6.post1
# vllm>=0.12.0
# transformers>=5.0.0rc0

# Optional: for development
# pytest>=7.0.0
# pre-commit>=4.5.0
# black>=23.0.0
# mypy>=1.0.0
